steamshortcuts cleanup
